Feature/nayeon - 모든 에러 처리 구현 완료

See merge request ai_track/class_07/ai_project/team07/7team_back!66

+++ src/schedulers/schedulers.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Post, Body, UseGuards, Get, Req, Param, Delete, ParseIntPipe } from "@nestjs/common";
+import { Controller, Post, Body, UseGuards, Get, Req, Param, Delete, ParseIntPipe, InternalServerErrorException, UnauthorizedException, ForbiddenException, BadRequestException } from "@nestjs/common";
 import { SchedulersService } from "./schedulers.service";
 import { JwtAuthGuard } from "src/auth/authentication/guards/jwt.guard";
 import {
@@ -18,28 +18,39 @@ export class SchedulersController {
   constructor(private readonly schedulersService: SchedulersService) {}
 
   //일정 생성 & 저장
-  @Post("ask")
+  @Post()
   @ApiOperation({ summary: "일정 생성" })
   async askGpt(
     @Req() req: any,
     @Body() createScheduleRequestDto: CreateScheduleRequestDto,
   ): Promise<CreateSchedulesResponseDto> {
+
     createScheduleRequestDto.userId = req.user.id;
-    return await this.schedulersService.askGpt(createScheduleRequestDto);
+    if(!req.user.id) throw new UnauthorizedException("로그인 되지 않은 사용자입니다.");
+
+    try {
+      return await this.schedulersService.askGpt(createScheduleRequestDto);
+    }
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
+    }
   }
 
   //일정 리스트 불러오기
   @Get()
   @ApiOperation({ summary: "일정 리스트 불러오기" })
   async getScheduleList(@Req() req: any): Promise<GetScheduleListResponseDto[]> {
+
     const userId = req.user.id;
+    if(!userId) throw new UnauthorizedException("로그인 되지 않은 사용자입니다.");
 
     try {
       const scheduleList = await this.schedulersService.getScheduleList(userId);
 
       return scheduleList;
-    } catch (error) {
-      throw new Error(error.message);
+    }
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
     }
   }
 
@@ -53,11 +64,11 @@ export class SchedulersController {
   ): Promise<GetSchedulesResponseDto> {
     try {
       const schedule = await this.schedulersService.getSchedule(schedulerId);
-      console.log(schedule);
 
       return schedule;
-    } catch (error) {
-      throw new Error(error.message);
+    }
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
     }
   }
 
@@ -71,20 +82,32 @@ export class SchedulersController {
     @Param() deleteScheduleRequestDto: DeleteScheduleRequestDto,
   ): Promise<MessageResponseDto> {
     const userId = req.user.id;
+    if(!userId || userId === undefined){
+      throw new UnauthorizedException("로그인 되지 않은 사용자입니다.");
+    }
+
+    const schedulerId = req.params.schedulerId;
+    if(!schedulerId || schedulerId == undefined){
+      throw new BadRequestException("유효하지 않은 요청입니다.")
+    }
 
     //본인 확인
     if (deleteScheduleRequestDto.userId !== userId) {
-      throw new Error("사용 권한이 없습니다.");
+      throw new ForbiddenException("삭제 권한이 없습니다.");
     }
 
     try {
       const schedulerId = +deleteScheduleRequestDto.schedulerId;
+      if(!schedulerId || schedulerId == undefined){
+        throw new BadRequestException("유효하지 않은 요청입니다.");
+      }
 
       const responseMessage = await this.schedulersService.deleteSchedule(schedulerId);
 
       return responseMessage;
-    } catch (error) {
-      throw new Error(error);
+    }
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
     }
   }
 }
