+++ src/users/users.controller.ts
@@ -11,6 +11,9 @@ import {
   UseInterceptors,
   UploadedFile,
   InternalServerErrorException,
+  UnauthorizedException,
+  ForbiddenException,
+  BadRequestException,
 } from "@nestjs/common";
 import { UsersService } from "./users.service";
 import { UsersRequestDto } from "./dto/users.request.dto";
@@ -23,6 +26,7 @@ import { MessageResponseDto } from "../common/dto/message.dto";
 import { User } from "@prisma/client";
 import { ApiFile } from "src/common/decorators/apiFile.decorator";
 import { FileInterceptor } from "@nestjs/platform-express";
+import { FORBIDDEN_MESSAGE } from "@nestjs/core/guards";
 
 @UseGuards(JwtAuthGuard)
 @ApiTags("user")
@@ -35,7 +39,6 @@ export class UsersController {
   @UseGuards(JwtAuthGuard)
   @ApiOperation({ summary: "현재 로그인한 사용자 정보 가져오기" })
   async getCurrentUser(@Req() req: any): Promise<User | null> {
-    console.log("req: ", req);
     const userId = req.user.id;
     try {
       const currentUserInfo = await this.usersService.getCurrentUser(userId);
@@ -51,26 +54,57 @@ export class UsersController {
   @ApiOperation({ summary: "마이페이지 이동" })
   @ApiResponse(GetUserInfoResponse)
   async getUserById(@Param() usersRequestDto: UsersRequestDto): Promise<myPageResponseDto> {
+    try {
     return this.usersService.getUserById(usersRequestDto.id);
+    }
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
+    }
   }
 
-  // 마이 프로필 업데이트 (프로필 사진(profilePath) | 유저네임(userName) | 자기소개(description) )
+  // 마이 프로필 업데이트 (유저네임(userName) | 자기소개(description) )
   @Patch(":id")
   @ApiOperation({ summary: "프로필 수정" })
   @ApiResponse(GetUserInfoResponse)
   async updateUserInfo(
     @Req() req: any,
     @Param() usersRequestDto: UsersRequestDto,
-    @Body("profilePath") profilePath: string | null,
     @Body("userName") username: string | null,
     @Body("description") description: string | null,
   ): Promise<myPageResponseDto> {
-    //본인 확인
-    if (usersRequestDto.id !== req.user.id) {
-      throw new Error("수정할 수 있는 권한 없음.");
-    }
+    if(!req.user.id) throw new UnauthorizedException("로그인 되지 않은 사용자입니다.");
 
-    return await this.usersService.updateUserInfo(usersRequestDto.id, profilePath, username, description);
+    try {
+      //본인 확인
+      if (usersRequestDto.id !== req.user.id) {
+        throw new ForbiddenException("수정할 수 있는 권한 없음.");
+      }
+
+      return await this.usersService.updateUserInfo(usersRequestDto.id, username, description);
+    }
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
+    }
+  }
+
+  //프로필 이미지 등록
+  @Post("Image")
+  @ApiFile()
+  @UseInterceptors(FileInterceptor("file"))
+  async uploadFile(@Req() req: any, @UploadedFile() file: Express.Multer.File): Promise<myPageResponseDto> {
+    const userId = req.user.id;
+    if(!userId) throw new UnauthorizedException("로그인 되지 않은 사용자입니다.");
+
+    if (!file || file == undefined){
+      throw new BadRequestException("유효한 이미지가 아닙니다.");
+    }
+    try {
+      const user = await this.usersService.uploadProfileImage(userId, file);
+      return user;
+    }
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
+    }
   }
 
   //계정 삭제
@@ -78,28 +112,21 @@ export class UsersController {
   @ApiOperation({ summary: "계정 삭제" })
   @ApiResponse(MessageResponse("계정을 성공적으로 삭제하였습니다."))
   async deleteUser(@Req() req: any, @Param() usersRequestDto: UsersRequestDto): Promise<MessageResponseDto> {
+    
+    if(!req.user.id) throw new UnauthorizedException("로그인 되지 않은 사용자입니다.");
+
+    try {
     //본인 확인
-    console.log("req.id: ", req.user.id);
-    console.log("userRequestDto.id :", usersRequestDto.id);
     if (req.user.id !== usersRequestDto.id) {
-      throw new Error("수정할 수 있는 권한 없음.");
+      throw new ForbiddenException("삭제할 수 있는 권한 없음.");
     }
 
     const result = await this.usersService.deleteUser(req.user.id);
 
     return result;
-  }
-
-  @Post("Image")
-  @ApiFile()
-  @UseInterceptors(FileInterceptor("file"))
-  async uploadFile(@Req() req: any, @UploadedFile() file: Express.Multer.File): Promise<myPageResponseDto> {
-    const userId = req.user.id;
-    console.log("userId: ", userId);
-    if (!file) {
-      throw new Error("File is not provided");
     }
-    const user = await this.usersService.uploadProfileImage(userId, file);
-    return user;
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
+    } 
   }
 }
